+++ resources/main.py
@@ -33,6 +33,9 @@ class OpenCoreLegacyPatcher:
         self.constants.recovery_status = utilities.check_recovery()
         self.computer = self.constants.computer
         self.constants.booted_oc_disk = utilities.find_disk_off_uuid(utilities.clean_device_path(self.computer.opencore_path))
+        if self.constants.computer.firmware_vendor:
+            if self.constants.computer.firmware_vendor != "Apple":
+                self.constants.host_is_hackintosh = True
         launcher_script = None
         launcher_binary = sys.executable
         if "python" in launcher_binary:
@@ -46,7 +49,6 @@ class OpenCoreLegacyPatcher:
         self.constants.unpack_thread.start()
 
         defaults.generate_defaults.probe(self.computer.real_model, True, self.constants)
-        global_settings.global_settings().convert_defaults_to_global_settings()
 
         if utilities.check_cli_args() is not None:
             print("- Detected arguments, switching to CLI mode")