+++ resources/validation.py
@@ -66,8 +66,6 @@ def validate(settings):
         host_os_float = float(f"{major_kernel}.{minor_kernel}")
         for patch_subject in patchset:
             for patch_core in patchset[patch_subject]:
-                if patch_core == "Nvidia Web Drivers":
-                    continue
                 patch_os_min_float = float(f'{patchset[patch_subject][patch_core]["OS Support"]["Minimum OS Support"]["OS Major"]}.{patchset[patch_subject][patch_core]["OS Support"]["Minimum OS Support"]["OS Minor"]}')
                 patch_os_max_float = float(f'{patchset[patch_subject][patch_core]["OS Support"]["Maximum OS Support"]["OS Major"]}.{patchset[patch_subject][patch_core]["OS Support"]["Maximum OS Support"]["OS Minor"]}')
                 if (host_os_float < patch_os_min_float or host_os_float > patch_os_max_float):