+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Downgrade AppleFSCompressionTypeZlib to 12.3.1 on pre-Sandy Bridge Macs
   - Resolves ZLib decompression kernel panics on 12.4 and newer
 - Resolve AppleGVACore crashing on MacBookPro11,3 in Monterey 12.4+
+- Add Nvidia Web Driver support for Maxwell and Pascal
+  - Currently running in OpenGL mode, [non-Metal issues](https://github.com/dortania/OpenCore-Legacy-Patcher/issues/108) applicable
 
 ## 0.4.5
 - Fix AutoPatcher.pkg download on releases